Imoprt: move ExportOCAF2 to own source files and move global functions to Tools class
